pre arrangements to implement ticket ordering logic

+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: dopereir <dopereir@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 18:53:26 by dopereir          #+#    #+#             */
-/*   Updated: 2025/03/01 05:10:44 by dopereir         ###   ########.fr       */
+/*   Updated: 2025/03/01 06:09:50 by dopereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ int	ft_atoi(const char *str)
 	return (sign * result);
 }
 
-long	get_current_time_ms(void)
+long long	get_current_time_ms(void)
 {
 	struct timeval	tv;
 
@@ -48,13 +48,13 @@ long	get_current_time_ms(void)
 void	print_message(t_data *data, const char *action)
 {
 	pthread_mutex_lock(data->print_message);
-	printf("%ld %d %s\n",
+	printf("%lld %d %s\n",
 		get_current_time_ms() - data->start_time_ms,
 		data->philo_id, action);
 	pthread_mutex_unlock(data->print_message);
 }
 
-long	get_elapsed_time(long start_time_ms)
+long long	get_elapsed_time(long start_time_ms)
 {
 	return (get_current_time_ms() - start_time_ms);
 }
